chore(cli): add OpenAI model flag to magicWand command
This commit adds a new flag to the `magicWand` command in the `cli` package, which allows the user to define what OpenAI model to use. The flag is defined in the `root` command and passed down to the `magicWand` command through its entry function. This change enables the usage of multiple OpenAI models in the `magicWand` command.

The `modelMap` variable has also been defined in the `root` command to map the user input to the respective OpenAI model constant value. Two values have been added to this map, `gpt35` and `gpt4`, representing two different OpenAI models.