feat(pkg/tzap): add ErrorTzap struct and methods
BREAKING CHANGE: This commit adds new functionality to the tzap package.

This commit adds a new struct, ErrorTzap, to handle errors in the tzap library. The ErrorTzap struct contains a pointer to a tzap object and an error.

The ErrorTzap struct has two methods. The first method, ErrorTzap, takes an error parameter and returns a pointer to an ErrorTzap object. The second method, HandleError, takes a function that accepts an *ErrorTzap parameter and returns a *Tzap object. If ErrorTzap.Err is not nil, the HandleError function will call the function passed as a parameter and pass it the ErrorTzap object. If the function returns nil, HandleError will panic with the error contained in ErrorTzap.Err.

The RequestTextToSpeech and RequestTextifySpeech methods have been updated to return an *ErrorTzap object. If an error occurs, the methods now call the ErrorTzap method and return the pointer to the ErrorTzap object.