feat(pkg/tzap): refactor onNew to AddContext
This commit renames the method onNew to AddContext in the tzap package. Also, it removes the addId method from the Tzap struct and moves it inside the AddContext method. The methods AddTzap, CloneTzap, and HijackTzap now use the new AddContext method instead of onNew method. Finally, the commit updates the tests in tzap_test.go to make use of the new AddContext method.

BREAKING CHANGE: The onNew method of the tzap package has been renamed to AddContext. The addId method has been removed from the Tzap struct and moved inside the AddContext method.