chore: Update version to v0.7.9 in Makefile and go.mod files
This commit updates the version to v0.7.9 in the following files:
- cli/Makefile
- cli/go.mod
- examples/go.mod
- pkg/connectors/googlevoiceconnector/go.mod
- pkg/connectors/openaiconnector/go.mod
- pkg/tzapconnect/go.mod

It also modifies the release target in the Makefile to run the 'test', 'build', and 'github-release' targets under the new 'releaseUpload' target. The 'release' target was changed to 'build', which only creates the binary and the ../release directory. The 'clean' target now also deletes the ../release directory.

These changes were made as part of updating the version of tzap to v0.7.9.