build: add gomodtidy target to Makefile
This commit adds the `gomodtidy` target to the `build` target in the Makefile, so that it runs before `go test` and building the CLI.

The `tzapWriteRelease` target has been moved to the `cli/Makefile`.

The `gh-upload` and `github-pre-release` targets have been simplified and moved to a separate release script.

Additionally, some redundant code in the `github-pre-release` target has been removed.

BREAKING-CHANGE: The `tzapWriteRelease` target has been moved from the root Makefile to `cli/Makefile`. This is a breaking change for anyone relying on this target in their build process.