fix: Update version to v0.7.11 in install.js
The previous version was v0.7.10, and this commit updates it to v0.7.11 in the install.js file.

Also, remove the .js files from the .gitignore file, and update the name of the CLI from "cli" to "tzap" in root.go. Lastly, remove github.copilot from the recommended extensions list in extensions.json.