fix(pkg/tzap): fix HandleError method signature
The HandleError method in the tzap package was meant to receive a callback function that returns an error value, but it was wrongly defined as a callback function that returns a *Tzap value instead. This caused panics when returning nil from the callback function. This commit fixes the HandleError method signature and adjusts the usage of it in tests.

BREAKING CHANGE: The signature of the HandleError callback function in the tzap package has changed from func(*tzap.ErrorTzap) *tzap.Tzap to func(*tzap.ErrorTzap) error, which may affect code that relies on this method.