Refactor(pkg/tzap): update ApplyTemplate function signature
The ApplyTemplate function in pkg/tzap/templates.go had its signature updated to receive a NamedTemplate with two parameters of types *Tzap and *Tzap, respectively. This change was made to enable the use of template functions that include an error parameter, which can be passed to the new ApplyErrorTemplate function.

Also, the ApplyErrorTemplate function was added to apply templates that can return errors. It receives a NamedTemplate with two parameters of types *Tzap and *ErrorTzap, respectively, and a function that returns an error. This function applies the template and applies the provided function to the resulting *ErrorTzap, handling any returned errors with a panic.