feat(cli): remove generate command and associated code
BREAKING CHANGE: The generate command in the CLI has been removed, along with the associated code in the cli/cmd/generate.go file. This command generated Tzap code, and is no longer necessary. This is a breaking change as any automation or scripts reliant on this command will need to be updated.