feat: add parameter to FetchTask in tzap/fetch-chat.go and applyChanges function in tzap/file2.go
BREAKING CHANGE: This change adds a new parameter to the FetchTask function in tzap/fetch-chat.go and applyChanges in tzap/file2.go, which requires changes to any code calling these functions. The added parameter is the filePath that will be fetched and edited. These changes also remove the `filepath` key from the `data` argument of FetchTask and applyChanges functions. The GetMD5Hash function in `tzap/file_extra.go` has been replaced with the `MD5Hash` function from the `util` package.