feat(pkg/tzapconnector): add localdbconnector to partial composite
This commit adds localdbconnector to the PartialComposite struct in tzapconnector's connector.go file. The purpose is to enable the use of the EmbeddingGenerator methods provided by this connector. The AddEmbeddingDocument, GetEmbeddingDocument, DeleteEmbeddingDocument, SearchWithEmbedding, and ListAllEmbeddingIds methods are included in PartialComposite.

BREAKING CHANGE: This commit does not include any breaking change.

# Explanation
The PartialComposite struct is used to create a partial implementation of the TGenerator interface. The addition of localdbconnector to the struct enables the use of the EmbeddingGenerator methods provided by this connector. These methods will be used to add, retrieve, and delete embedding documents, as well as to search for embeddings, and list them all.

This change does not introduce any breaking change to the codebase.