feat: add code generation and testing example
This commit adds a code generation and testing example to the `examples` directory.

The code generation example uses the `tzap` package to apply the `InspirationTemplate` to a list of files and adds a custom user message. The example also includes an optional command line argument to read from a file.

The testing example uses the `tzap` package to load files from a directory using the `LoadFileDir` method. The example also includes a prompt message for creating unit tests and adds a new template from an external package.

BREAKING CHANGE: The `MD5Rewrites` configuration option in the `TestingTask` configuration has been changed to `false`.