chore: Clean up npm package and Makefile
This commit cleans up the npm package and Makefile by removing the `bin` directory and updating the package version to `0.7.11-u`. It also adds `process.argv.slice(2)` to the `spawn` call in `npm/run.js` to pass any command line arguments to `tzap`. Finally, it pipes `process.stdin` to `tzap.stdin` to allow `tzap` to accept input from the console.