chore(cli): Update tzap version from v0.7.11 to v0.7.12
This commit updates the tzap version from v0.7.11 to v0.7.12 in the Makefile and cli/Makefile. It also updates the PREVIOUS_VERSION and VERSION variables in cli/Makefile to v0.7.12. Additionally it adds a new task tzapPrepareRelease in Makefile to prepare for the release and fetches the git tags after creating a release. The commit also updates the description in cli/cmd/ghrelease.go to be more descriptive about the required output of the JSON template.