chore: update version number in npm package
Updated the version number of the npm package from 0.7.11 to 0.7.12 in the package.json and install.js files. Also changed a console.log statement to write to stdout and a console.error statement to write to stderr in the run.js file. This is not a breaking change and does not affect any functionality, only updates the package version number.