Chore(ignores): Adding ignore files for better control over embedded files
Preparing for the ability to steer what files are embedded with .gitignore, .tzapignore, .tzapinclude. This commit adds the necessary files to the repository, namely .gitignore, .tzapignore, and .tzapinclude.

The changes in .gitignore ignore the "out/" and ".redis-data/" folders. The changes in .tzapignore exclude certain folders and files like "vendor", "npm", and ".vscode". Additionally, this file adds an example of how to ignore test files with "*_test.go". The new file .tzapinclude is intended to be used in conjunction with the .gitignore and .tzapignore files and specifies what files to include in the build.

Note that the library used for .tzapignore is not 100% compliant with how .gitignore is used. Therefore, use it with caution.