feat(pkg/util)! now returns relative path
Adds the ability to return a list of files under a directory, where the paths of the files are relative to the directory. This is useful for when we want to use the files in a different directory than the one they were found in. The previous implementation returned a list of absolute paths.

BREAKING CHANGE: The return value of `ListFilesInDir` function has changed from `[]string` to `([]string, error)`. The function now returns an error if the provided directory does not exist.

This change also adds a test for the `ListFilesInDir` function.