feat(cli/cmd/util): add ability to steer what files are embedded with .gitignore, .tzapignore, .tzapinclude
This commit introduces a function `GetNonExcludedFiles` which takes a list of file names, filters them based on patterns specified in `.gitignore`, `.tzapignore`, and `.tzapinclude` files, and returns the filtered list.

The `ReadFilterPatternFiles` function reads the contents of `.gitignore`, `.tzapignore`, or `.tzapinclude` files and returns the list of patterns to be used for filtering the files.

The `FilterWithExcludePattern`, `FilterFileSize` and `FilterWithIncludePattern` functions filter the files based on the patterns specified in `.gitignore`, `.tzapignore`, and `.tzapinclude` files.

Unit tests are also added to test the new functionality.

BREAKING CHANGE: function signature may change for function callers, but won't affect the service.