refactor: renamed template package to workflow in tzap
The templates.go and templates_test.go files have been deleted, and the workflow.go and workflow_test.go files have been added in the tzap package. This is a renaming of the template package to workflow package for better clarity and consistency in the codebase. The ApplyTemplateP, ApplyTemplateFN and their related functions have been renamed to ApplyWorkflowP, ApplyWorkflowFN and their related functions in the new workflow package.