feat(util/file, util/stdin): add new functions for writing files and confirming input prompts
This commit adds two new functions to the util/file package and util/stdin package respectively.

The `MkdirPAndWriteFile` function added to the util/file package creates a directory for the file if it doesn't exist and writes the edited content to the file.

The `ConfirmPrompt` function added to the util/stdin package prompts for a yes or no confirmation from the user and returns `true` in case of 'y'/'Y' and `false` in case of 'n'/'N'.

These functions have been added to improve the user experience and file writing functionality of the tool.

BREAKING CHANGE: The `output` package has been deleted as it was redundant and not used anywhere in the codebase.