refactor(pkg/tzap): rename function `ApplyTemplateFN` to `ApplyWorkflowFN`
The `ApplyTemplateFN` function was renamed to `ApplyWorkflowFN` to better reflect its purpose, which is to apply a workflow function to each component of a `Tzap` tree. The name `ApplyWorkflowFN` better describes this function since a workflow function is applied to each node. Updates were also made to function calls in `controlflow.go` and `error_test.go` to reflect the function name change.

BREAKING CHANGE: `ApplyTemplateFN` function was renamed to `ApplyWorkflowFN` function. This change may break existing code that relies on the previous function name for its function calls.