feat(cli): add flags to embeddingPrompt command
This commit adds two optional flags to the `embeddingPrompt` command. The `-k` or `--embeds` flag is used to specify the number of embeddings to use for prompt generation. The `-n` or `--searchsize` flag is used to specify the number of embeddings to include in the search space before filtering out the matches with inspiration files. These flags have default values of 10 and 15 respectively.