chore: update command help wording for ghrelease and gitcommit commands
Update the help messages of `ghrelease` and `gitcommit` commands to use the term "Template" instead of "Workflow" when referring to the JSON template for the release and Git commit message, respectively. Also, update an example in the `ghrelease` command to use proper headings in the template. This will make the command help messages more clear and consistent with the terminology used in the CLI.