feat(pkg/tzap): Add Reduce method to controlflow.go
This commit adds a new method `Reduce` to the `controlflow.go` file in the `pkg/tzap` package. The `Reduce` method takes a function that reduces the children of a `Tzap` object and returns a new `Tzap` object representing the result of the reduction. This method can be used to perform operations on nested `Tzap` objects, such as summing the values of all children or finding the maximum value.