feat(logging): add support for logging Dot language files
This commit adds support for logging Dot language files as a new log type in the `filelog` package. Now, the `LogData` function in the package checks if the log type is `DotLog` and if it is, saves the data as a string with the `.dot` extension. Also, a new `writeStringToFile` function is added to save the Dot log file.

The `GenerateGraphvizDotFile` function in `log-graphviz-builder.go` has been modified to use the `filelog.LogData` function to save the Dot file instead of writing it directly to the filesystem.

Finally, the `HandleShutdown` function in `log-graphviz.go` is replaced with a new method associated with the `tzap.Tzap` struct that handles the shutdown logic. The new method, `HandleShutdown`, receives a `Tzap` instance and is intended to be called on shutdown to generate the Dot graph and flush the logs.

BREAKING CHANGE: `HandleShutdown` function is replaced with a new method associated with the `tzap.Tzap` struct, which receives the instance and replaces the global `GlobalTzaps` slice. This change also breaks the `main.go` file in the `translategotots` package, which was calling the global `HandleShutdown` function. This function call has been removed, and users of `tzap` will need to update the way they handle shutdown logic.