feat(cli): add flag to read from file instead of prompt
This commit adds a new flag to the root command (`--file` or `-f`) in order to read input from a file instead of a prompt. This flag is used in the `semanticgitcommit` command to load staged commit diffs from file.

BREAKING CHANGE: The `--file` flag changes the behavior of the root command and therefore can be considered a breaking change. The `--show-diff` flag was also added to the `semanticgitcommit` command to show git diff.

Other changes include removing some unnecessary print statements in `embed/query.go` and `tzap/file.go`, and adding a stack trace to errors in `tzap/error.go`. The `magicWand` command was also updated to fix a typo in its description.