feat(cli): add support for loading OpenAI API key from environment variable or .env file
This commit adds support for loading the OpenAI API key from an environment variable or a `.env` file. If the key is not found in any of these sources, an error message is returned with instructions on how to add the key to either the environment variable or `.env` file.

The `loadAPIKey` function in `pkg/tzapconnect/loadapikey.go` has been modified to remove the code that tried to get the API key from a flag. Also, the function signature has been changed to accept the key as an argument.

The changes to `loadAPIKey` function required the modification of the `rootCmd` in `cli/cmd/root.go` to pass the key as an argument to the `loadAPIKey` function. However, the modification has been skipped for the `init` and `help` subcommands.