feat(request): modify request completion to use thread instead of messages
BREAKING CHANGE: This commit changes the functionality of the fetchChatResponse function in the request completion package to use the GetThread function instead of the GetMessages function. This may require changes in dependent packages that were using the fetchChatResponse function.

This commit also adds a new thread package and implements the GetThread function to return all the messages in the thread of the provided message. The function TruncateToMaxWords is updated to work with the thread, truncating to the maximum word count while taking the messages across the entire thread.

The StoreThread and LoadThread functions are also added to store and load threads to and from files.