feat: add temperature flag for OpenAI GPT-3 generation
BREAKING CHANGE: default temperature for OpenAI now set to 1.0

Add a new `--temperature` flag to the `tzapcli` command. This flag allows the user to set the temperature for OpenAI GPT-3 generation. The default temperature is now 1.0.

This change requires an update to the `Configuration` struct and the `fetchChatResponse` function in the `OpenaiTgenerator` implementation. The `Temperature` field has been added to the `Configuration` struct, and the `Temperature` value is now obtained from the context's configuration in the `fetchChatResponse` function. This update also introduces a breaking change, as the default temperature for OpenAI models is now 1.0 instead of not being set.