feat: add topbar and menu to documentation page
Added a topbar to the documentation page which includes a logo and a hamburger menu for mobile devices. Also added a menu which gets expanded upon clicking the hamburger menu, and added links to the different sections of the page. This is a new feature added to the codebase.

BREAKING CHANGE: Removed margin-bottom from the title of the page.