Update TZAP website files and add metadata for social media sharing
Update index.html to add metadata for social media sharing, including title, description, image, URL, and type. Also added a new favicon.ico and logo_whiteonblack.png, as well as a LICENSE file.

Deleted install.sh script.

Changed the header logo to include an image and text, and made it responsive.

Changed the topbar-content div to a nav element and added a hamburger menu for small screens.

Added a new section with a demo, and a comparison image between GPT completion and existing code.

Added a new section for the TZAP community and a link to the Discord server.

Made various style updates to improve website aesthetics and responsiveness.