chore: Add favicon files for various sizes and platforms
Add favicon files for different sizes and platforms such as android-chrome and apple-touch-icon. These files were added in the docs/favicons directory. Also, an additional favicon file was added in the top-level favicon directory.

This is a chore as it does not add any new feature or functionality, but rather improves the appearance of the website and makes it more recognizable in various platforms and devices.