#Summary This commit adds badges for Discord, Official Page, NPM, and Twitter Follow to the README.md file. It also updates the introductory text and adds some more information to the existing text.
#Changes Made
- Added badges for Discord, Official Page, NPM, and Twitter Follow to the README.md file.
- Updated the introductory text to include the Tzap Dev Tools.
- Added more information to the existing text of the README.md file.

#Motivation
Adding badges and updating the text of the README.md file will make it easier for users to join Tzap's community, visit the official page, and follow Tzap on social media. It will also provide more information about what the Tzap Dev Tools are and how they work, making it easier for potential users to learn about and use the tool.