feat: Bump version to v0.7.17 and update related files
Bump the version of Tzap to v0.7.17 in the Makefiles and package.json files. Also update the version in the install.js and run.js files. This includes updating the VERSION variable in the cli/Makefile and the Version property in the cli/cmd/root.go file. Remove the public directory from the npm/Makefile as it is unnecessary.

BREAKING CHANGE: None