docs: Update index.html with clearer prompt for feature building
Changed the prompt for "Feature building" section in index.html to provide a clearer idea about the functionality that needs to be implemented. The previous example was vague and did not specify the type of subscription to be implemented with stripe. This commit updates the prompt to read "How would we implement a new paid subscription model using stripe in our graphql backend?".

Also, updated the comparison.png image in docs and public directories. The diff shows a binary file difference between the old and new images.