docs: Update index.html with new demo and comparison images
Added a new section to the index.html file containing a demo of the tzap prompt interface, with an accompanying image. Also added a new comparison image to the public directory and updated the reference to it in the index.html file. Removed the old demo and comparison images, which were located in the npm/public directory.