feat(codegen): refactor makecode function and remove redundant go and ts implementations
This commit refactors the `MakeCode` function in the `codegeneration` package to be more generic and allows it to generate code in any language. The function now takes the input and output file extensions as arguments and generates the output file by replacing the input file extension with the output file extension. The commit also removes the redundant `MakeCodeGO` and `MakeCodeTS` functions and their implementations.

BREAKING CHANGE: The `MakeCode` function signature has changed and now requires the language, input and output file extensions, mission and task arguments to be explicitly passed in. The removed functions `MakeCodeGO` and `MakeCodeTS` are no longer available. Code generators invoking these functions need to be updated to use the new `MakeCode` signature.